Extract `MaybeFade` to its own component for easier further extractions later if used across components
