+++ packages/components/src/drop-zone/animators/index.ts
@@ -1,0 +1,1 @@
+export { MaybeFade } from './maybe-fade';
